requirements.txt
cat > requirements.txt << 'EOF'
# IDM Generative System — Streamlit Cloud dependencies
numpy>=1.26.0
scipy>=1.11.0
soundfile>=0.12.0
librosa>=0.10.0
fastapi>=0.110.0
pydantic>=2.0.0
python-dotenv>=1.0.0
qdrant-client>=1.8.0
openai>=1.0.0
streamlit>=1.32.0
EOF
